# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Property {
  id: number;
  slug: string;
  title: string;
  location: string;
  city: string;
  country: string;
  propertyType: string;
  listingType: string;
  status: string;
  priceLabel: string;
  bedrooms: number;
  bathrooms: number;
  area: number;
  areaUnit: string;
  image: string;
  imageAlt?: string;
  featured: boolean;
  shortDescription?: string;
}

export type PropertyDetail = Property & ({
  description: string;
  images: string[];
  amenities: string[];
  developer: string;
  completionStatus: string;
  nearby?: string[];
  /** @nullable */
  videoUrl?: string | null;
  /** @nullable */
  brochureUrl?: string | null;
});

export interface EnquiryInput {
  /** @minLength 2 */
  name: string;
  email: string;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  country?: string | null;
  /** @minLength 5 */
  message: string;
  /** @nullable */
  propertySlug?: string | null;
  enquiryType?: string;
  /** @nullable */
  preferredContact?: string | null;
}

export interface EnquiryReceipt {
  id: number;
  message: string;
}

export type ListPropertiesParams = {
city?: string;
country?: string;
propertyType?: string;
listingType?: string;
/**
 * @minimum 0
 */
minBedrooms?: number;
featured?: boolean;
q?: string;
};