# lib/api-zod/src/generated/types/enquiryInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface EnquiryInput {
  /** @minLength 2 */
  name: string;
  email: string;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  country?: string | null;
  /** @minLength 5 */
  message: string;
  /** @nullable */
  propertySlug?: string | null;
  enquiryType?: string;
  /** @nullable */
  preferredContact?: string | null;
}