# lib/api-zod/src/generated/types/property.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Property {
  id: number;
  slug: string;
  title: string;
  location: string;
  city: string;
  country: string;
  propertyType: string;
  listingType: string;
  status: string;
  priceLabel: string;
  bedrooms: number;
  bathrooms: number;
  area: number;
  areaUnit: string;
  image: string;
  imageAlt?: string;
  featured: boolean;
  shortDescription?: string;
}