# lib/api-zod/src/generated/types/propertyDetail.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Property } from './property';

export type PropertyDetail = Property & ({
  description: string;
  images: string[];
  amenities: string[];
  developer: string;
  completionStatus: string;
  nearby?: string[];
  /** @nullable */
  videoUrl?: string | null;
  /** @nullable */
  brochureUrl?: string | null;
});